requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
requests==2.32.4
python-dotenv==1.0.1
googletrans==4.0.0rc1
torch==2.7.1
torchvision==0.22.1
pillow==11.2.1
numpy==2.1.3
opencv-python==4.10.0.84